feat(MainModelHandler): AdvertisementPolicy와 테스트의 위치를 ui로 옮기고 main model의 구조 변경
