+++ app/src/main/java/woowacourse/movie/model/main/AdvertisementMapper.kt
@@ -1,0 +1,17 @@
+package woowacourse.movie.model.main
+
+import woowacourse.movie.R
+import woowacourse.movie.advertisement.Advertisement
+
+object AdvertisementMapper {
+    fun Advertisement.toUiModel(): AdvertisementUiModel {
+        return AdvertisementUiModel(
+            id = this.id,
+            link = this.link,
+            image = R.drawable.woowa_advertisement,
+        )
+    }
+
+    fun List<Advertisement>.toUiAdvertisements(): List<AdvertisementUiModel> =
+        map { it.toUiModel() }
+}
