+++ app/src/main/java/woowacourse/movie/model/main/MainData.kt
@@ -1,5 +1,7 @@
 package woowacourse.movie.model.main
 
+import woowacourse.movie.ui.main.adapter.MainViewType
+
 sealed class MainData {
     abstract val mainViewType: MainViewType
     abstract val id: Long
