+++ app/src/main/java/woowacourse/movie/ui/main/adapter/listview/MainViewHolder.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.ui.main.adapter.listview
 
 import android.view.View
 import woowacourse.movie.model.main.MainData
-import woowacourse.movie.model.main.MainViewType
+import woowacourse.movie.ui.main.adapter.MainViewType
 
 sealed class MainViewHolder(val view: View) {
     abstract fun onBind(data: MainData)
