+++ app/src/main/java/woowacourse/movie/ui/main/adapter/recyclerview/MainViewHolder.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.ui.main.adapter.recyclerview
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.model.main.MainData
-import woowacourse.movie.model.main.MainViewType
+import woowacourse.movie.ui.main.adapter.MainViewType
 
 sealed class MainViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
     abstract fun onBind(data: MainData)
